Update location of `ReceiptController` and rename to `ReceiptsController`
